Update index.html
The problem was not Discord rendering it weirdly, it was discord just not updating the preview with the new meta. it cached the old meta or something and I'm not sure I have a way to force Discord to update it. Masking the link with a redirect like through a link shortener is one workaround which tricks it into thinking it's a new link and so it does read the new meta.